UPDATED sx.observables: unified .retryWith implementation

+++ libs/sx-common/src/main/java/sx/rx/Observables.kt
@@ -115,7 +115,7 @@ fun <T> Observable<T>.behave(supplier: CompositeDisposableSupplier? = null): Obs
  */
 fun <T> Observable<T>.retryWith(
         // TODO: Int.MAX_VALUE doesn't seem to work for this implementation (freezes), while it works with Completable which is essentially the same implementation
-        count: Int,
+        count: Short,
         action: (retry: Long, error: Throwable) -> Observable<Long> = { _, _ -> Observable.just(0) })
         : Observable<T> {
